requirements.txt
astroid==2.2.5
colorama==0.4.1
cycler==0.10.0
isort==4.3.21
joblib==0.13.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.17.1
opencv-python==4.1.0.25
pandas==0.25.1
Pillow==6.1.0
plotly==4.1.1
pylint==2.3.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
retrying==1.3.3
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
